Phase 3: Sheet Management, QA, Collaboration & UI Modernization

+++ StingTools/Temp/TemplateExtCommands.cs
@@ -296,7 +296,7 @@ namespace StingTools.Temp
                     try
                     {
                         ViewSchedule vs = ViewSchedule.CreateMaterialTakeoff(
-                            doc, new ElementId(bic));
+                            doc, new ElementId((long)bic));
                         vs.Name = name;
                         created++;
                     }
